MiNubeIA/requirements.txt
google-genai>=0.3.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pvporcupine>=3.0.0
pyaudio>=0.2.13
